+++ my-webxr-app/src/components/createTicks.tsx
